+++ app/angular/src/client/preview/angular-beta/RendererService.test.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
+import { Parameters } from '../types-6-0';
 import { RendererService } from './RendererService';
 
 jest.mock('@angular/platform-browser-dynamic');
@@ -181,6 +181,35 @@ describe('RendererService', () => {
       });
 
       expect(countDestroy).toEqual(1);
+    });
+
+    describe('bootstrap module options', () => {
+      async function setupComponentWithWhitespace(bootstrapModuleOptions: unknown) {
+        await rendererService.render({
+          storyFnAngular: {
+            template: '<div>   </div>',
+            props: {},
+          },
+          forced: false,
+          parameters: {
+            bootstrapModuleOptions,
+          } as Parameters,
+        });
+      }
+
+      it('should preserve whitespaces', async () => {
+        await setupComponentWithWhitespace({ preserveWhitespaces: true });
+        expect(document.body.getElementsByTagName('storybook-wrapper')[0].innerHTML).toBe(
+          '<div>   </div>'
+        );
+      });
+
+      it('should remove whitespaces', async () => {
+        await setupComponentWithWhitespace({ preserveWhitespaces: false });
+        expect(document.body.getElementsByTagName('storybook-wrapper')[0].innerHTML).toBe(
+          '<div></div>'
+        );
+      });
     });
   });
 });