+++ lib/cli/src/repro-generators/configs.ts
@@ -116,13 +116,21 @@ export const angular: Parameters = baseAngular;
 export const web_components: Parameters = {
   framework: 'web-components',
   name: 'web_components',
-  version: 'latest',
+  version: '2',
   generator: fromDeps('lit-element'),
 };
 
 export const web_components_typescript: Parameters = {
   ...web_components,
   name: 'web_components_typescript',
+  typescript: true,
+};
+
+export const web_components_lit2: Parameters = {
+  ...web_components,
+  version: 'next',
+  name: 'web_components_lit2',
+  generator: fromDeps('lit'),
   typescript: true,
 };