+++ lib/core-common/src/types.ts
@@ -1,5 +1,5 @@
 import type ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-import type { PluginOptions } from 'react-docgen-typescript-plugin';
+import type { PluginOptions } from '@storybook/react-docgen-typescript-plugin';
 import { Configuration, Stats } from 'webpack';
 import { TransformOptions } from '@babel/core';
 import { Router } from 'express';
@@ -144,6 +144,7 @@ export interface CLIOptions {
   debugWebpack?: boolean;
   webpackStatsJson?: string | boolean;
   outputDir?: string;
+  modern?: boolean;
 }
 
 export interface BuilderOptions {
@@ -236,6 +237,7 @@ export interface StorybookConfig {
    */
   features?: {
     postcss?: boolean;
+    buildStoriesJson?: boolean;
   };
   /**
    * Tells Storybook where to find stories.