+++ lib/manager-webpack4/src/presets/manager-preset.ts
@@ -1,6 +1,6 @@
 import { Configuration } from 'webpack';
 import { loadManagerOrAddonsFile, ManagerWebpackOptions, Options } from '@storybook/core-common';
-import createDevConfig from '../manager/manager-webpack.config';
+import createDevConfig from '../manager-webpack.config';
 
 export async function managerWebpack(
   _: Configuration,
@@ -11,10 +11,12 @@ export async function managerWebpack(
 
 export async function managerEntries(
   installedAddons: string[],
-  options: { managerEntry: string; configDir: string }
+  options: { managerEntry: string; configDir: string; modern?: boolean }
 ): Promise<string[]> {
   const { managerEntry = '@storybook/core-client/dist/esm/manager' } = options;
-  const entries = [require.resolve('../globals/polyfills')];
+  const entries = options.modern
+    ? []
+    : [require.resolve('@storybook/core-client/dist/esm/globals/polyfills')];
 
   if (installedAddons && installedAddons.length) {
     entries.push(...installedAddons);